+++ apps/api/src/shared/kernel/value-object.base.ts
@@ -1,8 +1,4 @@
-interface ValueObjectProps {
-  [index: string]: unknown;
-}
-
-export abstract class ValueObject<T extends ValueObjectProps> {
+export abstract class ValueObject<T> {
   public readonly props: T;
 
   constructor(props: T) {